perf: only resync on non-final entity states (#185)
The refetch due to stale data is only very useful for very old webhooks (protected by timestamp upsert) or when there are multiple webhooks in a short span of time.

For the latter, if the entity is already in a final state, there is a low chance of another webhook coming in for the same entity.

To avoid unnecessarily spamming the Stripe API, this PR introduces skipping the refetch if entities are already in final states.

Note that new webhooks, i.e. changed metadata will still be processed, it just avoids the refetch via API.